Neural Network Practice
I tried to work with tensorflow to be able to create the algorithm but I was stuck even after watching several videos. By the end of it I kinda figured out how to set up a Neural Network with tensorflow, I just don't really know how to use it to sort out data like that. In the file ethan_nn.py is just a neural network with commented out code on how I set it up, I played around with the code and tried to understand what it was doing.